requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web interface
streamlit>=1.25.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better performance
# tensorboard>=2.13.0
# wandb>=0.15.0
